+++ crates/usvg/src/writer.rs
@@ -113,18 +113,17 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
         xml.start_svg_element(EId::Filter);
         xml.write_id_attribute(filter.id(), opt);
         xml.write_rect_attrs(filter.rect);
-        xml.write_units(AId::FilterUnits, filter.units, Units::ObjectBoundingBox);
         xml.write_units(
-            AId::PrimitiveUnits,
-            filter.primitive_units,
+            AId::FilterUnits,
             Units::UserSpaceOnUse,
+            Units::ObjectBoundingBox,
         );
 
         for fe in &filter.primitives {
             match fe.kind {
                 filter::Kind::DropShadow(ref shadow) => {
                     xml.start_svg_element(EId::FeDropShadow);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &shadow.input);
                     xml.write_attribute_fmt(
                         AId::StdDeviation.to_str(),
@@ -139,7 +138,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::GaussianBlur(ref blur) => {
                     xml.start_svg_element(EId::FeGaussianBlur);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &blur.input);
                     xml.write_attribute_fmt(
                         AId::StdDeviation.to_str(),
@@ -150,7 +149,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Offset(ref offset) => {
                     xml.start_svg_element(EId::FeOffset);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &offset.input);
                     xml.write_svg_attribute(AId::Dx, &offset.dx);
                     xml.write_svg_attribute(AId::Dy, &offset.dy);
@@ -159,7 +158,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Blend(ref blend) => {
                     xml.start_svg_element(EId::FeBlend);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &blend.input1);
                     xml.write_filter_input(AId::In2, &blend.input2);
                     xml.write_svg_attribute(
@@ -188,7 +187,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Flood(ref flood) => {
                     xml.start_svg_element(EId::FeFlood);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_color(AId::FloodColor, flood.color);
                     xml.write_svg_attribute(AId::FloodOpacity, &flood.opacity.get());
                     xml.write_svg_attribute(AId::Result, &fe.result);
@@ -196,7 +195,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Composite(ref composite) => {
                     xml.start_svg_element(EId::FeComposite);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &composite.input1);
                     xml.write_filter_input(AId::In2, &composite.input2);
                     xml.write_svg_attribute(
@@ -225,7 +224,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Merge(ref merge) => {
                     xml.start_svg_element(EId::FeMerge);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_svg_attribute(AId::Result, &fe.result);
                     for input in &merge.inputs {
                         xml.start_svg_element(EId::FeMergeNode);
@@ -237,14 +236,14 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Tile(ref tile) => {
                     xml.start_svg_element(EId::FeTile);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &tile.input);
                     xml.write_svg_attribute(AId::Result, &fe.result);
                     xml.end_element();
                 }
                 filter::Kind::Image(ref img) => {
                     xml.start_svg_element(EId::FeImage);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_aspect(img.aspect);
                     xml.write_svg_attribute(
                         AId::ImageRendering,
@@ -273,7 +272,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::ComponentTransfer(ref transfer) => {
                     xml.start_svg_element(EId::FeComponentTransfer);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &transfer.input);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
@@ -286,7 +285,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::ColorMatrix(ref matrix) => {
                     xml.start_svg_element(EId::FeColorMatrix);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &matrix.input);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
@@ -312,7 +311,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::ConvolveMatrix(ref matrix) => {
                     xml.start_svg_element(EId::FeConvolveMatrix);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &matrix.input);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
@@ -346,7 +345,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Morphology(ref morphology) => {
                     xml.start_svg_element(EId::FeMorphology);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &morphology.input);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
@@ -370,7 +369,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::DisplacementMap(ref map) => {
                     xml.start_svg_element(EId::FeDisplacementMap);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_filter_input(AId::In, &map.input1);
                     xml.write_filter_input(AId::In2, &map.input2);
                     xml.write_svg_attribute(AId::Result, &fe.result);
@@ -395,7 +394,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::Turbulence(ref turbulence) => {
                     xml.start_svg_element(EId::FeTurbulence);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
                     xml.write_attribute_fmt(
@@ -427,7 +426,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::DiffuseLighting(ref light) => {
                     xml.start_svg_element(EId::FeDiffuseLighting);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
                     xml.write_svg_attribute(AId::SurfaceScale, &light.surface_scale);
@@ -439,7 +438,7 @@ fn write_filters(tree: &Tree, opt: &XmlOptions, xml: &mut XmlWriter) {
                 }
                 filter::Kind::SpecularLighting(ref light) => {
                     xml.start_svg_element(EId::FeSpecularLighting);
-                    xml.write_filter_primitive_attrs(fe);
+                    xml.write_filter_primitive_attrs(filter.rect(), fe);
                     xml.write_svg_attribute(AId::Result, &fe.result);
 
                     xml.write_svg_attribute(AId::SurfaceScale, &light.surface_scale);
@@ -846,7 +845,7 @@ trait XmlWriterExt {
     fn write_numbers(&mut self, aid: AId, list: &[f32]);
     fn write_image_data(&mut self, kind: &ImageKind);
     fn write_filter_input(&mut self, id: AId, input: &filter::Input);
-    fn write_filter_primitive_attrs(&mut self, fe: &filter::Primitive);
+    fn write_filter_primitive_attrs(&mut self, parent_rect: NonZeroRect, fe: &filter::Primitive);
     fn write_filter_transfer_function(&mut self, eid: EId, fe: &filter::TransferFunction);
 }
 
@@ -934,6 +933,7 @@ impl XmlWriterExt for XmlWriter {
         });
     }
 
+    // TODO: simplify
     fn write_units(&mut self, id: AId, units: Units, def: Units) {
         if units != def {
             self.write_attribute(
@@ -1010,18 +1010,18 @@ impl XmlWriterExt for XmlWriter {
         );
     }
 
-    fn write_filter_primitive_attrs(&mut self, fe: &filter::Primitive) {
-        if let Some(n) = fe.x {
-            self.write_svg_attribute(AId::X, &n);
+    fn write_filter_primitive_attrs(&mut self, parent_rect: NonZeroRect, fe: &filter::Primitive) {
+        if parent_rect.x() != fe.rect().x() {
+            self.write_svg_attribute(AId::X, &fe.rect().x());
         }
-        if let Some(n) = fe.y {
-            self.write_svg_attribute(AId::Y, &n);
+        if parent_rect.y() != fe.rect().y() {
+            self.write_svg_attribute(AId::Y, &fe.rect().y());
         }
-        if let Some(n) = fe.width {
-            self.write_svg_attribute(AId::Width, &n);
+        if parent_rect.width() != fe.rect().width() {
+            self.write_svg_attribute(AId::Width, &fe.rect().width());
         }
-        if let Some(n) = fe.height {
-            self.write_svg_attribute(AId::Height, &n);
+        if parent_rect.height() != fe.rect().height() {
+            self.write_svg_attribute(AId::Height, &fe.rect().height());
         }
 
         self.write_attribute(